pyproject.toml
[tool.poetry]
name = "logpirate"
version = "0.1.0"
description = "Convert BurpSuite logs to HAR"
authors = ["Peter Kasza <peter.kasza@itinsight.hu>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
python-dateutil = "^2.8.2"
lxml = "^4.9.2"

[tool.poetry.scripts]
logpirate = 'logpirate:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
